models for paper trade account
paper trade orders

float to int

postloginresponse, lmt price to string, alert rules, paper trade, Market order type, lmtPirce, trade token, api get options quotes

api get options quotes

fundamentals

active gainers

+++ openapi/model_add_alert_request.go
@@ -17,7 +17,7 @@ type AddAlertRequest struct {
 	ExchangeCode string `json:"exchangeCode,omitempty"`
 	RegionId string `json:"regionId,omitempty"`
 	ShowCode string `json:"showCode,omitempty"`
-	TickerId float32 `json:"tickerId,omitempty"`
+	TickerId int32 `json:"tickerId,omitempty"`
 	TickerName string `json:"tickerName,omitempty"`
 	TickerSymbol string `json:"tickerSymbol,omitempty"`
 	TickerType string `json:"tickerType,omitempty"`
